Feat : adding cmm page
